+++ api/feeder-api/src/main/java/cz/inqool/dl4dh/feeder/kramerius/dto/FeedResponseDto.java
@@ -1,0 +1,13 @@
+package cz.inqool.dl4dh.feeder.kramerius.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+import java.util.Map;
+
+@Getter
+@Setter
+public class FeedResponseDto {
+    private List<Map<String, Object>> data;
+}